requirements.txt
streamlit
folium
streamlit_folium
pandas
langchain
python-dotenv
requests
FAISS
pandas
openai
langchain
faiss-cpu
langchain-community